tests: remove duplicates and only test benchmarks once by default
